library/requirements.txt
django==3.2.0
djangorestframework==3.13.0